Simplify taglib interactions further
