+++ osu.Game/Rulesets/Edit/Checks/CheckAudioInVideo.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using System.IO;
 using osu.Framework.Logging;
 using osu.Game.Beatmaps;
-using osu.Game.IO.FileAbstraction;
 using osu.Game.Rulesets.Edit.Checks.Components;
 using osu.Game.Storyboards;
 using osu.Game.Utils;
@@ -62,7 +61,7 @@ namespace osu.Game.Rulesets.Edit.Checks
                 {
                     // We use TagLib here for platform invariance; BASS cannot detect audio presence on Linux.
                     using (Stream data = context.WorkingBeatmap.GetStream(storagePath))
-                    using (File tagFile = TagLibUtils.CreateFile(new StreamFileAbstraction(filename, data)))
+                    using (File tagFile = TagLibUtils.GetTagLibFile(filename, data))
                     {
                         if (tagFile.Properties.AudioChannels == 0)
                             continue;
